src/requirements.txt
flask
werkzeug
Pillow
numpy